added authorization for community and post

+++ app/Http/Controllers/Backend/CommunityController.php
@@ -68,6 +68,7 @@ class CommunityController extends Controller
      */
     public function edit(Community $community)
     {
+        $this->authorize('update', $community);
         return Inertia::render('Communities/Edit', compact('community'));
     }
 
@@ -80,6 +81,8 @@ class CommunityController extends Controller
      */
     public function update(CommunityUpdateRequest $request, Community $community)
     {
+        $this->authorize('update', $community);
+
         $community->update($request->validated());
 
         return to_route('communities.index')->with('message', 'Community updated successfully');
@@ -93,6 +96,8 @@ class CommunityController extends Controller
      */
     public function destroy(Community $community)
     {
+        $this->authorize('delete', $community);
+
         $community->delete();
         return back()->with('message', 'Community deleted successfully');
     }
